Explicitly form table urns for regexes
